pyproject.toml
[project]
name = "yattmcp"
version = "0.1.0"
description = "Yet Another TickTick MCP"
readme = "README.md"
authors = [
    { name = "Maxwell Koo", email = "30420+mjkoo@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "typer>=0.21.1",
]

[project.scripts]
yattmcp = "yattmcp.cli:app"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
